+++ CHANGELOG.md
@@ -19,11 +19,13 @@ and this project adheres to
 
 ### Fixed
 
+- DHCP lease's `expired` field incorrect time format ([#2692]).
 - Incomplete DNS upstreams validation ([#2674]).
 - Wrong parsing of DHCP options of the `ip` type ([#2688]).
 
 [#2674]: https://github.com/AdguardTeam/AdGuardHome/issues/2674
 [#2688]: https://github.com/AdguardTeam/AdGuardHome/issues/2688
+[#2692]: https://github.com/AdguardTeam/AdGuardHome/issues/2692